schedule, negotiation mapper 삭제 #52
