+++ src/main/java/com/tsp/new_tsp_front/api/model/domain/negotiation/FrontNegotiationEntity.java
@@ -12,6 +12,8 @@ import javax.validation.constraints.Email;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
+import java.util.List;
+import java.util.stream.Collectors;
 
 import static javax.persistence.FetchType.LAZY;
 import static javax.persistence.GenerationType.IDENTITY;
@@ -26,7 +28,7 @@ import static javax.persistence.GenerationType.IDENTITY;
 @Table(name = "tsp_model_negotiation")
 public class FrontNegotiationEntity extends NewCommonMappedClass {
     @Transient
-    private Integer rnum;
+    private Integer rowNum;
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
@@ -71,4 +73,31 @@ public class FrontNegotiationEntity extends NewCommonMappedClass {
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "model_idx", referencedColumnName = "idx", insertable = false, updatable = false)
     private FrontModelEntity frontModelEntity;
+
+    public static FrontNegotiationDTO toDto(FrontNegotiationEntity entity) {
+        if (entity == null) return null;
+        return FrontNegotiationDTO.builder()
+                .idx(entity.getIdx())
+                .rowNum(entity.getRowNum())
+                .modelIdx(entity.getModelIdx())
+                .modelKorName(entity.getModelKorName())
+                .modelNegotiationDesc(entity.getModelNegotiationDesc())
+                .modelNegotiationDate(entity.getModelNegotiationDate())
+                .name(entity.getName())
+                .email(entity.getEmail())
+                .phone(entity.getPhone())
+                .visible(entity.getVisible())
+                .creator(entity.getCreator())
+                .createTime(entity.getCreateTime())
+                .updater(entity.getUpdater())
+                .updateTime(entity.getUpdateTime())
+                .build();
+    }
+
+    public List<FrontNegotiationDTO> toDtoList(List<FrontNegotiationEntity> entityList) {
+        if (entityList == null) return null;
+        return entityList.stream()
+                .map(FrontNegotiationEntity::toDto)
+                .collect(Collectors.toList());
+    }
 }
