+++ src/main/java/com/tsp/new_tsp_front/api/model/domain/schedule/FrontScheduleEntity.java
@@ -11,6 +11,9 @@ import javax.persistence.*;
 import javax.validation.constraints.NotEmpty;
 import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.stream.Collectors;
 
 import static javax.persistence.FetchType.EAGER;
 import static javax.persistence.FetchType.LAZY;
@@ -26,7 +29,7 @@ import static javax.persistence.GenerationType.IDENTITY;
 @Table(name = "tsp_model_schedule")
 public class FrontScheduleEntity extends NewCommonMappedClass {
     @Transient
-    private Integer rnum;
+    private Integer rowNum;
 
     @Id
     @GeneratedValue(strategy = IDENTITY)
@@ -54,4 +57,26 @@ public class FrontScheduleEntity extends NewCommonMappedClass {
     @ManyToOne(fetch = LAZY)
     @JoinColumn(name = "model_idx", referencedColumnName = "idx", insertable = false, updatable = false)
     private FrontModelEntity frontModelEntity;
+
+    public static FrontScheduleDTO toDto(FrontScheduleEntity entity) {
+        if (entity == null) return null;
+        return FrontScheduleDTO.builder().idx(entity.getIdx())
+                .rowNum(entity.getRowNum())
+                .modelIdx(entity.getModelIdx())
+                .modelSchedule(entity.getModelSchedule())
+                .modelScheduleTime(entity.getModelScheduleTime())
+                .visible(entity.getVisible())
+                .creator(entity.getCreator())
+                .createTime(entity.getCreateTime())
+                .updater(entity.getUpdater())
+                .updateTime(entity.getUpdateTime())
+                .build();
+    }
+
+    public static List<FrontScheduleDTO> toDtoList(List<FrontScheduleEntity> entityList) {
+        if (entityList == null) return null;
+        return entityList.stream()
+                .map(FrontScheduleEntity::toDto)
+                .collect(Collectors.toList());
+    }
 }
