+++ src/main/java/com/tsp/new_tsp_front/api/model/schedule/service/impl/FrontScheduleJpaRepository.java
@@ -17,7 +17,6 @@ import java.util.Map;
 
 import static com.tsp.new_tsp_front.api.model.domain.QFrontModelEntity.*;
 import static com.tsp.new_tsp_front.api.model.domain.schedule.QFrontScheduleEntity.*;
-import static com.tsp.new_tsp_front.api.model.schedule.service.impl.FrontScheduleMapper.INSTANCE;
 import static com.tsp.new_tsp_front.common.utils.StringUtil.getInt;
 import static com.tsp.new_tsp_front.common.utils.StringUtil.getString;
 import static java.time.LocalDate.now;
@@ -86,9 +85,9 @@ public class FrontScheduleJpaRepository {
                 .fetch();
 
         scheduleList.forEach(list -> scheduleList.get(scheduleList.indexOf(list))
-                .setRnum(getInt(scheduleMap.get("startPage"), 1) * (getInt(scheduleMap.get("size"), 1)) - (2 - scheduleList.indexOf(list))));
+                .setRowNum(getInt(scheduleMap.get("startPage"), 1) * (getInt(scheduleMap.get("size"), 1)) - (2 - scheduleList.indexOf(list))));
 
-        return INSTANCE.toDtoList(scheduleList);
+        return FrontScheduleEntity.toDtoList(scheduleList);
     }
 
     /**
